# src/AspNetCoreSample.WebApi/OpenAPI/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "AspNetCoreSample WebApi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Keycloak/FetchUser": {
      "get": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザー情報を取得",
        "operationId": "Keycloak_FetchUser",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "ユーザー名",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/CreateUser": {
      "post": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザーを作成する",
        "operationId": "Keycloak_CreateUser",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "description": "ユーザー情報",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/UpdateUser": {
      "put": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザー情報を更新する",
        "operationId": "Keycloak_UpdateUser",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "description": "ユーザー情報",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/UpdateUserByUsername": {
      "put": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザー情報を更新する\nユーザー名で更新する",
        "operationId": "Keycloak_UpdateUserByUsername",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserByUsernameInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/ChangePassword": {
      "put": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザーのパスワードを変更する",
        "operationId": "Keycloak_ChangePassword",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "description": "パスワード情報",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/ChangePasswordByUsername": {
      "put": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザーのパスワードを変更する",
        "operationId": "Keycloak_ChangePasswordByUsername",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "description": "パスワード情報",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordByUsernameInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/ResetPasswordByEmail": {
      "put": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザーのパスワードをリセットするためのメールを送信",
        "operationId": "Keycloak_ResetPasswordByEmail",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordByEmailInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Keycloak/ResetPasswordByEmailByUsername": {
      "put": {
        "tags": [
          "Keycloak"
        ],
        "summary": "ユーザー名でユーザーのパスワードをリセットするためのメールを送信",
        "operationId": "Keycloak_ResetPasswordByEmailByUsername",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "input",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordByEmailByUsernameInput"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/Keycloak/FetchRoles": {
      "get": {
        "tags": [
          "Keycloak"
        ],
        "summary": "Realmに登録されているロール一覧を取得",
        "operationId": "Keycloak_FetchRoles",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "firstName",
          "lastName",
          "email",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ChangePasswordInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "password"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ChangePasswordByUsernameInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "description": "Bearer {token}",
        "in": "header",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}